Session: simplify UPnP setup and avoid race condition
Checking for a mapping availability and setting it up in two
different moments may lead to race conditions as other procsses may
register the same mapping before lbrynet does.

Moreover, the return value of upnp.addportmapping() was not checked,
thus leading lbrynet to believe that the mapping was succesful even if
it was not.

Starting with miniupnpc 2.0.x, we have support for the
upnp.addanyportmapping() primitive which instructs the server to reserve
the first available external port if the requested one is already taken.

If no port is avaialable the upnp call will fail.

This new primitive therefore avoids the need to "test and set" the
mapping thus eliminating the racy behaviour.

Simplify the UPnP setup by using upnp.addanyportmapping() and check
its return value to ensure everything went fine.

Signed-off-by: Antonio Quartulli <antonio@mandelbit.com>