INSTALL.md: add `libleveldb-dev` to required packages
With the provided instructions in `INSTALL.md`,
the `make` command may result in errors:
```
    plyvel/_plyvel.cpp:589:10: fatal error: leveldb/db.h: No such file or directory
      589 | #include "leveldb/db.h"
          |          ^~~~~~~~~~~~~~
    compilation terminated.
    error: command 'x86_64-linux-gnu-gcc' failed with exit status 1

...
Failed to build plyvel prometheus-client pylru pyyaml.
```

The required header files are contained in the `libleveldb-dev`
package in Ubuntu 18.04 and 20.04 at least.

After installing this package the compilation proceeds successfully.

In some cases `make` may have to be run twice to compile everything
without problems.