Daemon.jsonrpc_claim_search: `page_size` is set to `DEFAULT_PAGE_SIZE` if none entered
Currently with `lbrynet claim search`
when `--page_size` is specified, it will use the minimum
of this value or 50. So, if it's `--page_size=1000` it will still use 50.

If `--page_size` is omitted, it will use the minimum
of `DEFAULT_PAGE_SIZE` (normally 20) or 50, so it will use 20.

So, instead of using `min` we simply remove this function.
In this case, it will use simply use the value of `page_size`,
or if this is missing, it will use `DEFAULT_PAGE_SIZE`,
which should always be available.

By using a large `page_size`, this will allow us to fit
as many elements as possible in the claim search.

However, there is still a separate bug, lbry-sdk issue #3365,
which limits the total results to 1000, no matter the `page_size`.