Daemon.jsonrpc_get: add `download_collection` option for handling collections
Normally, if a particular claim is in fact a collection (playlist),
it will not be downloaded, as it is not a stream.
```
lbrynet get collection-claim
```

Will result in
```
{
  "error": "Claim is not a stream."
}
```

To resolve the collection, and download the claims that it contains
we can use the option `--download_collection`.
```
lbrynet get collection-claim --download_collection
```

A collection can have an arbitrary number of items, so to limit
the number of claims to download we can use `--max_claims`.
```
lbrynet get collection-claim --download_collection --max_claims=4
```

To reverse the order of the claims in the collection
use `--reverse_collection`. This can be used to get the newest
items in the collection.
```
lbrynet get collection-claim --download_collection --max_claims=10 --reverse_collection
```