Daemon.jsonrpc_file_summary: new method to print a summary of claims
This allows printing a list of all claim streams that were downloaded
to the system.
The list is printed to the terminal or to a specific file.

It accepts some parameters to control the information that is printed.
```
lbrynet file summary --blobs --show_channel --title --stream_type --path
lbrynet file summary --show=incomplete --start=10 --end=40
lbrynet file summary --sort=claim_name --reverse --sep=' ;.;'
```

The `--file` option writes the list of claims to a file
which then can be shared with other users of LBRY in order
to download the same claims and contribute to seeding that content.
```
lbrynet file summary --channel=@SomeChannel --file=summary.txt --fdate
```

By default it will print the date of the claim which is based on the `'claim_height'`,
when the claim was registered in the blockchain, the `'claim_id'`, the `'claim_name'`,
and whether the media is present or not in the download directory.
```
 1/42; 20200610_10:23:37-0500; b231714456ee832daeba4b8356803e7591126dff; "07-S"; no-media
 2/42; 20200610_10:27:06-0500; 31700ff11f900429d742f2f137ba25393bdb3b0a; "09-S"; media
 3/42; 20200609_23:14:47-0500; 70dfefa510ca6eee7023a2a927e34d385b5a18bd; "04-S"; no-media
```